viewer: Don't recompute the canvas size in all image drawing!!
On a fast desktop, recomputing the CSS width/height of the canvas can
take anywhere between 0.3ms and 1.5ms, or more, whereas a cached imaged
is rendered in less than 1ms. This is needless slowdown, that will be
even more visible on lower end devices.

Move to record the canvas size in the resize handler, and cache it, so
that in drawImage() this is no longer needed.