viewer: Fix a very long-standing bug in the canvas positioning
This bothere me for ages, and I wasn't able to understand when it
triggers. While trying to optimise the canvas resize, I realised the
navbar can resize on its own, without triggering the window resize
event, which led to a long conversation with Claude, which ended up with
a very awesome solution. Although, as the changelog says, in some cases
it's a bit disrupting change in the canvas size.

In any case, this allows removing the manual tracking of navbar height
and the repositioning of the #main div.