viewer: Move to absolute, pixel-based panning
This took three times the amount of time I thought it will, because a
very simply thing stumped me (the previous code's moving from [-1, 1] to
[0, 2], which was adding a centering offset basically), but in the
process I had to go line-by-line, or rather variable by variable, in
drawImage.

And man, there was a lot of tech debt. Including a formula which, in
three separate steps, was basically doing:

  a = (b - c) / 2 - (b / 2)

Instead of the direct: a = -c / 2. The end effect is that while this is
a large diff, its removing twice the number of lines that it adds,
rouhgly (!), and simplifies significantly the drawImage function, which
now is simply:

- compute target image size
- thanks to the canvas coordinate system being (0, 0) = center of
  screen, for centering the image, just negate its scaled and zoomed
  size, and halve it; this directly gives the offsets (one variable
  instead of 3 or 4)
- to compute the final drawin position, take the centered offsets and
  simply add, unmodified, the panning offsets, with the only addition
  that the panning offsets should be bounded by the panning limits

This makes drawImage clearer. I'll do a round of renames for even better
clarity.