viewer: fix image nav group positioning on iOS
Commit 0f7d49b1 changed to make the main div relative, to fix the video
positioning via creating a new stacking context. This was all good for
video, but on iOS/iPadOS, but not Safari desktop, this broke the image
nav group, since it was also using position: absolute and not fixed.
Since that navbar has to be on that position, always, change to fixed.