viewer: Add back overscroll-behavior: none on the body element
I had this at one point, and it was fine, then I removed it and didn't
see any change, then much later strange bugs appeared. After adding it
back, _and_ resetting the iPad simulator, it's back to normal.

Claude says this is needed on both html and body (and that's how it
originally recommended), because apparently there are differences on
desktop and mobile on which exact element is "the primary scroll
container". So let's fix this, forever, since 20 more bytes on every
page is a million times better than extra whitespace at the top of the
page, as if reload failed but the page was stuck there.