viewer: Adjust mouse wheel zooming to account for trackpads
In the default setup, without taking into account the event deltaY
values, the trackpad scrolls too fast. But using the actual values is
too fast for the mouse and too slow for the trackpad, so use a variation
that tries to balance both. Still a bit too fast zooming, but showing
off a bit is not that bad :)