Navbar/breadcrumbs: fix wrapping/resize on long paths
This was annoying for a really long time, and finally got to fix it. Not
very clean though, on image viewer it'd be best to show just the folder
name for my collection model, but an improvement.

Additionally, enable the breadcrumbs at a size lower (Bootstrap small vs
medium). Before, small would have basically guaranteed ugly wrapping,
hence it was disabled.