Breadcrumbs: customise for the image viewer
I'm pretty pleases with how this ended, as it gives a progressively
better context based on the screen size: first folder only (at small),
then folder and image (at medium), and then top level folder, folder and
image (at large). This doesn't work well if the top-level folder has
information that the folder name doesn't, but I think it's still better
than always showing everything, as the folder name should be relatively
representative (of the context, not of the image of course).