Convert camerastats to a module!
Finally, after lots of trying, one working version. And setup such that
the other plotly modules can easily be integrated, and still have just
one bundle for plotly-pages, so that the client can only download it
once.

The gain is, at least for this single plot, a decrease in half
(4.5MiB→2.2MiB) of total page resources. Will see once all plot types
are added, but still, this is a non-trivial gain.