fancybox: upgrade to v5 and convert to a module
This is less successful than previous ones, but mostly due to
defficiencies in the new v5 API (or my understanding of it). Filled
upstream bug asking for clarification
(https://github.com/fancyapps/ui/issues/747), in the meantime it is what
it is - but moving forward since this completes the module and bundler
move, and now cleanup can be done.