Start of searching behaviour changes: folders->list
The grid for folders doesn't make much sense, so until it does,
constrain folders by default to list mode.

This requires though that folders don't overwrite the view mode anymore,
since that would force images to list. This is partially correct, as
right now we'd still want to mark "folder (list)", but not overwrite
image browse to image list. Sigh.