Major change: drop the fancybox interface
I wanted to do this for a very long time, but the lack of first movie
playin (fixed a while ago) and second zooming, which was recently fixed,
prevented me. Now that both roadblocks have been removed, I did a simple
test on how this would look. And it looks much better, or at least
straightforward, as the view mode (for image browsing) will be 1:1 with
the list of images in the browse mode. Plus, navigating back from the view
mode to the browse mode is easy, at least with keyboard.

This also eliminates the bugs with the new version of fancybox
(technically, `@fancyapps/ui` now), so it's a double plus.