Rework the 'dev' cabal flag to enable more testing
Currently, the flag defaults to false to ensure that it doesn't get
automatically enabled during the release. However, this prevents
conditional (#ifdef'ed) exports, which are IMHO the best way to
conditionally export functions for testing (lacking Rust's "cfg(test)").
Rework this by making the production binaries unbuildable when this flag is
set, and change it to be set by default, and make the test suite
conditional on it being set. This way the test suite and library can use
the mentioned conditional exports (to be added in the future).