Add some hard-gained View tests
The initial AI test suggestions were quite a bit off, as the basic
infrastructure to insert mock images in the repository was missing, so a
big chunk of this commit is that. The actual test is there, but the
framework is way more important, as a very wide number of handler tests
are very rudimentary today without custom images.