Migrate from TH-based to Generic-based Data.Store instances
I'm not convinced, since I can't easily reproduce the old pains with TH extra
recompilation, but based on the type of the data types (products, not sums),
Generic-based should be similar performance, and the
serialisation/deserialisation is not in the hot path. The goal is to keep TH
only in the Yesod parts, and remove it from the "backend" code.

Especially the really base types should need no TH.