viewer: show an indeterminate progress bar during image load
This will give some interactivity, on slow connections, so that the interface
feels responsive.

It would be best to do this only after, let's say, 50ms and not immediately, so
that on fast connections, there's no flickering of the cursor or progress bar.
So the feature is not yet fully baked.