viewer: Rework how exif information is built
Apparently, javascript doesn't have a built-in safe, escape function?! So let's
move everything to the backend, and ship the fully-built HTML for the element.

This is a bit strange, but it simplifies things a lot: it keeps all HTML
rendering in Hamlet, and removes the need for actual JSON encoding of the Exif
structure. I'll remove that once I'm fully happy with the output.